remove --discover-sibling, fix bug with user moving the point
Discovery of siblings is not needed, because at any point we can
reference the first child of the node and check it's position.

Likewise --teleport-stack only needs two references, not an entire
list of it, because when we teleport we only check the "real" parent
against the parent from --teleport-stack.